+++ website/src/routes/schedules/index.lazy.tsx
@@ -2,14 +2,16 @@ import { createLazyFileRoute } from "@tanstack/react-router";
 import styled from "styled-components";
 import Slideshow from "./-components/Slideshow";
 import MovieBox from "./-components/MovieBox";
-import { useSchedules } from "./-hooks/useSchedules";
+import { useTransSchedules } from "./-hooks/useTransSchedules";
 import { Suspense } from "react";
+import { useSchedules } from "../../hooks/services/useSchedules";
 
 export const Route = createLazyFileRoute("/schedules/")({
   component: Index,
 });
 function Index() {
-  const { schedules } = useSchedules("10");
+  const { schedules } = useSchedules("");
+
   return (
     <SchedulesContainer>
       <Title>上映スケジュール</Title>
